refactor: UI 개발을 위한 mock을 하나의 파일로 모음
