--- app/note/_common/mock.ts
@@ -1,57 +1,0 @@
-export const categorys = [
-  {
-    categorySlug: "2023",
-    category: "2023 모음",
-  },
-  {
-    categorySlug: "blog",
-    category: "블로그 개발기",
-  },
-  {
-    categorySlug: "react",
-    category: "리액트",
-  },
-  {
-    categorySlug: "2022",
-    category: "2022 모음",
-  },
-  {
-    categorySlug: "backend",
-    category: "벡엔드",
-  },
-];
-
-export const notes = [
-  {
-    categorySlug: "blog",
-    category: "블로그 개발기",
-    titleSlug: "notion-api",
-    title: "Notion API 연동하기",
-    intro: "Notion API를 사용해서 블로그를 개발하는 방법을 알려드립니다.",
-    createdAt: "2022-10-20",
-  },
-  {
-    categorySlug: "2023",
-    category: "2023 모음",
-    titleSlug: "frontend-build",
-    title: "프론트엔드 개발 환경",
-    intro: "프론트엔드 개발 환경에 대한 전반적인 내용을 정리했습니다.",
-    createdAt: "2023-10-20",
-  },
-  {
-    categorySlug: "blog",
-    category: "블로그 개발기",
-    titleSlug: "notion-api",
-    title: "Notion API 연동하기 조금 많이 긴 제목이라면 어떻게할까요?",
-    intro: "Notion API를 사용해서 블로그를 개발하는 방법을 알려드립니다.",
-    createdAt: "2022-10-20",
-  },
-  {
-    categorySlug: "2023",
-    category: "2023 모음",
-    titleSlug: "frontend-build",
-    title: "프론트엔드 개발 환경",
-    intro: "프론트엔드 개발 환경에 대한 전반적인 내용을 정리했습니다.",
-    createdAt: "2023-10-20",
-  },
-];
